[tune] move force_on_current_node to ml_utils
