+++ doc/source/tune/api_docs/trainable.rst
@@ -362,8 +362,6 @@ Utilities
 
 .. autofunction:: ray.tune.utils.validate_save_restore
 
-.. autofunction:: ray.tune.utils.force_on_current_node
-
 
 .. _tune-ddp-doc:
