+++ python/ray/tune/examples/pbt_convnet_function_example.py
@@ -141,7 +141,7 @@ if __name__ == "__main__":
         # happens on the server. So we wrap `test_best_model` in a Ray task.
         # We have to make sure it gets executed on the same node that
         # ``tune.run`` is called on.
-        from ray.tune.utils.util import force_on_current_node
+        from ray.util.ml_utils.node import force_on_current_node
 
         remote_fn = force_on_current_node(ray.remote(test_best_model))
         ray.get(remote_fn.remote(analysis))
